split schema.py into three files
